# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: M4Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chats:
    get:
      operationId: getChats
      tags: [chats]
      summary: Get all chats for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Chats list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatItem"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create new chat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"
      responses:
        "201":
          description: Chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get chat by ID
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatDetail"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete/leave chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /chats/{chatId}/members:
    get:
      operationId: getChatMembers
      tags: [chats]
      summary: Get chat members
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMember"
    post:
      operationId: addChatMember
      tags: [chats]
      summary: Add member to chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
      responses:
        "201":
          description: Member added
  /chats/{chatId}/members/{userId}:
    delete:
      operationId: removeChatMember
      tags: [chats]
      summary: Remove member from chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
  /chats/{chatId}/invite:
    post:
      operationId: generateInviteLink
      tags: [chats]
      summary: Generate invite link
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite link
          content:
            application/json:
              schema:
                type: object
                properties:
                  inviteLink:
                    type: string
  /chats/join/{inviteLink}:
    post:
      operationId: joinByInvite
      tags: [chats]
      summary: Join chat by invite link
      security:
        - bearerAuth: []
      parameters:
        - name: inviteLink
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  # Messages
...
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload file
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                chatId:
                  type: integer
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileItem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Chat:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [private, group, channel]
        photo:
          type: string
        inviteLink:
          type: string
        pinnedMessageId:
          type: integer
        isArchived:
          type: boolean
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
    ChatItem:
      allOf:
        - $ref: "#/components/schemas/Chat"
        - type: object
          properties:
            role:
              type: string
            unreadCount:
              type: integer
            lastMessage:
              $ref: "#/components/schemas/MessagePreview"
    ChatDetail:
      allOf:
        - $ref: "#/components/schemas/Chat"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/ChatMember"
    CreateChatRequest:
      type: object
      required: [type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [private, group, channel]
        participantIds:
          type: array
          items:
            type: integer
    ChatMember:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [creator, admin, member, restricted]
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - role
...
    MessagePreview:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        messageType:
          type: string
        senderId:
          type: integer
        senderName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    FileItem:
      type: object
      properties:
        id:
          type: integer
        fileKey:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url