# lib/api-zod/src/generated/types/chatItem.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * M4Chat API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Chat } from './chat';
import type { MessagePreview } from './messagePreview';

export type ChatItem = Chat & {
  role?: string;
  unreadCount?: number;
  lastMessage?: MessagePreview;
};