# lib/api-zod/src/generated/types/fileItem.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * M4Chat API specification
 * OpenAPI spec version: 0.2.0
 */

export interface FileItem {
  id: number;
  fileKey?: string;
  fileName?: string;
  mimeType?: string;
  size?: number;
  url: string;
  createdAt?: Date;
}