# lib/api-zod/src/generated/types/messagePreview.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * M4Chat API specification
 * OpenAPI spec version: 0.2.0
 */

export interface MessagePreview {
  id?: number;
  content?: string;
  messageType?: string;
  senderId?: number;
  senderName?: string;
  createdAt?: Date;
}